Finish solution to Day 14
Required a total rewrite. I spent ages trying to solve part 2 by making
as much fuel as possible, then converting the surplus chemicals back to
ore, then trying again. But it got stuck in cycles where it was only
adding 1 fuel at a time, and took too long.

I had to go on Reddit for hints where people mentioned allowing the
stock to become negative then correcting it after. That breakthrough
allowed me to solve the problem - I wouldn't have thought of it,
although it's obvious now (and a generalisation of what I was doing
before)

The idea of doing a power of 10 binary search at the start is also
from Reddit.

First time I've had to look up hints for a question, which sucks,
but I already spent many hours on this!