Ugly solution to Day 15 Part 1
This solution explores the whole maze first, because I was curious,
then uses Djikstra to find the fastest route.

As the maze is basically a tree (does not contain cycles) we didn't
need to do both steps. We could have either used the backtracking
algorithm, counting the steps as we go; or used Djikstra while
interactively asking the robot to find walls. I'm hoping doing
both will save me work in Part 2.